requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bandit==1.7.4
base58==2.1.1
bitarray==2.5.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cytoolz==0.12.0
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.4.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.8.3
flake8==4.0.1
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
lru-dict==1.1.8
matplotlib-inline==0.1.3
mccabe==0.6.1
multiaddr==0.0.9
multidict==6.0.2
nest-asyncio==1.5.5
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
parso==0.8.3
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycryptodome==3.15.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rlp==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.3.0
stevedore==4.0.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
urllib3==1.26.10
varint==1.0.2
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
yarl==1.7.2
